[CONTINUUM-1886] fix NPE
git-svn-id: https://svn.eu.apache.org/repos/asf/continuum/trunk@698083 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/DefaultContinuum.java
@@ -802,16 +802,20 @@ public class DefaultContinuum
             Project project = getProjectWithBuilds( projectId );
 
             List<ContinuumReleaseResult> releaseResults = releaseResultDao.getContinuumReleaseResultsByProject( projectId );
-            
+
             try
             {
                 for ( ContinuumReleaseResult releaseResult : releaseResults )
                 {
                     releaseResultDao.removeContinuumReleaseResult( releaseResult );
                 }
-                
+
                 File releaseOutputDirectory = configurationService.getReleaseOutputDirectory( project.getProjectGroup().getId() );
-                FileUtils.deleteDirectory( releaseOutputDirectory );
+
+                if ( releaseOutputDirectory != null )
+                {
+                    FileUtils.deleteDirectory( releaseOutputDirectory );
+                }
             }
             catch ( ContinuumStoreException e )
             {
